vaccine_system_backend/requirements.txt

fastapi
uvicorn
sqlalchemy
databases[sqlite]
alembic
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
